# Conflicts:
#	README.md
#	h2/src/main/org/h2/command/Command.java
#	h2/src/main/org/h2/command/Parser.java
#	h2/src/main/org/h2/command/ddl/AlterTableAddConstraint.java
#	h2/src/main/org/h2/command/ddl/AlterTableAlterColumn.java
#	h2/src/main/org/h2/command/ddl/AlterTableRename.java
#	h2/src/main/org/h2/command/ddl/CreateDomain.java
#	h2/src/main/org/h2/command/ddl/CreateRole.java
#	h2/src/main/org/h2/command/ddl/CreateTable.java
#	h2/src/main/org/h2/command/ddl/CreateUser.java
#	h2/src/main/org/h2/command/ddl/DropSchema.java
#	h2/src/main/org/h2/command/dml/BackupCommand.java
#	h2/src/main/org/h2/command/dml/Delete.java
#	h2/src/main/org/h2/command/dml/Insert.java
#	h2/src/main/org/h2/command/dml/SetClauseList.java
#	h2/src/main/org/h2/command/dml/Update.java
#	h2/src/main/org/h2/command/query/Select.java
#	h2/src/main/org/h2/engine/ConnectionInfo.java
#	h2/src/main/org/h2/engine/Database.java
#	h2/src/main/org/h2/engine/Engine.java
#	h2/src/main/org/h2/engine/Right.java
#	h2/src/main/org/h2/engine/Role.java
#	h2/src/main/org/h2/engine/UndoLog.java
#	h2/src/main/org/h2/engine/UndoLogRecord.java
#	h2/src/main/org/h2/engine/User.java
#	h2/src/main/org/h2/expression/Expression.java
#	h2/src/main/org/h2/expression/ExpressionColumn.java
#	h2/src/main/org/h2/expression/aggregate/Aggregate.java
#	h2/src/main/org/h2/index/Index.java
#	h2/src/main/org/h2/jdbc/JdbcPreparedStatement.java
#	h2/src/main/org/h2/message/TraceObject.java
#	h2/src/main/org/h2/mvstore/Page.java
#	h2/src/main/org/h2/pagestore/Page.java
#	h2/src/main/org/h2/pagestore/PageFreeList.java
#	h2/src/main/org/h2/pagestore/PageLog.java
#	h2/src/main/org/h2/pagestore/PageOutputStream.java
#	h2/src/main/org/h2/pagestore/PageStore.java
#	h2/src/main/org/h2/pagestore/PageStreamData.java
#	h2/src/main/org/h2/pagestore/PageStreamTrunk.java
#	h2/src/main/org/h2/pagestore/db/PageBtree.java
#	h2/src/main/org/h2/pagestore/db/PageBtreeCursor.java
#	h2/src/main/org/h2/pagestore/db/PageBtreeIndex.java
#	h2/src/main/org/h2/pagestore/db/PageBtreeLeaf.java
#	h2/src/main/org/h2/pagestore/db/PageBtreeNode.java
#	h2/src/main/org/h2/pagestore/db/PageData.java
#	h2/src/main/org/h2/pagestore/db/PageDataIndex.java
#	h2/src/main/org/h2/pagestore/db/PageDataLeaf.java
#	h2/src/main/org/h2/pagestore/db/PageIndex.java
#	h2/src/main/org/h2/pagestore/db/PageStoreTable.java
#	h2/src/main/org/h2/pagestore/db/ScanIndex.java
#	h2/src/main/org/h2/result/RowList.java
#	h2/src/main/org/h2/server/TcpServer.java
#	h2/src/main/org/h2/table/Table.java
#	h2/src/main/org/h2/value/Value.java

+++ h2/src/main/org/h2/mvstore/type/ObjectDataType.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2020 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * Copyright 2004-2021 H2 Group. Multiple-Licensed under the MPL 2.0,
  * and the EPL 1.0 (https://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
@@ -1592,7 +1592,7 @@ public class ObjectDataType extends BasicDataType<Object> {
             int size = data.length * 2;
             // adjust the average size
             // using an exponential moving average
-            averageSize = (size + 15 * averageSize) / 16;
+            averageSize = (int) ((size + 15L * averageSize) / 16);
             buff.put((byte) TYPE_SERIALIZED_OBJECT).putVarInt(data.length)
                     .put(data);
         }
@@ -1609,7 +1609,7 @@ public class ObjectDataType extends BasicDataType<Object> {
             int size = data.length * 2;
             // adjust the average size
             // using an exponential moving average
-            averageSize = (size + 15 * averageSize) / 16;
+            averageSize = (int) ((size + 15L * averageSize) / 16);
             buff.get(data);
             return deserialize(data);
         }
