+++ h2/src/test/org/h2/test/db/TestBigDb.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2020 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * Copyright 2004-2021 H2 Group. Multiple-Licensed under the MPL 2.0,
  * and the EPL 1.0 (https://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
@@ -99,7 +99,7 @@ public class TestBigDb extends TestDb {
                     long t = System.nanoTime();
                     if (t - time > TimeUnit.SECONDS.toNanos(1)) {
                         time = t;
-                        int free = Utils.getMemoryFree();
+                        long free = Utils.getMemoryFree();
                         println("i: " + i + " free: " + free + " used: " + Utils.getMemoryUsed());
                     }
                 }
